Inline from_inline_const into its sole call site
